post merge fixes

+++ src/Projectile.cc
@@ -1,8 +1,10 @@
 #include "Projectile.h"
+
 #include <iostream>
 
 Projectile::Projectile(sf::Vector2f const &pos, double rotation, double velocity, double damage)
-    :Entity(std::string{"projectile"}, pos), velocity{velocity}, damage{damage}, texture{TextureManager::instance()->getTexture("AR_bullet.png")}
+    : Entity(std::string{"projectile"}, pos), velocity{velocity}, damage{damage},
+      texture{TextureManager::instance()->getTexture("AR_bullet.png")}
 {
     sf::Sprite::setTexture(*texture);
     sf::Sprite::setRotation(rotation);
@@ -19,7 +21,9 @@ void Projectile::onCollision(std::string const &other)
 
 void Projectile::move()
 {
-    sf::Vector2f dir {std::cos((sf::Sprite::getRotation() + 90) * M_PI / 180.0f), std::sin((sf::Sprite::getRotation() + 90)  * M_PI / 180.0f)}; 
+    sf::Vector2f dir{
+        static_cast<float>(std::cos((sf::Sprite::getRotation() + 90) * M_PI / 180.0f)),
+        static_cast<float>(std::sin((sf::Sprite::getRotation() + 90) * M_PI / 180.0f))};
 
     sf::Sprite::Transformable::move(-dir.x * velocity, -dir.y * velocity);
 }
